requirements.txt
# project requirements
pandas
googlesearch-python
python-dotenv
loguru
pydantic

# dev requirements
pre-commit
black
flake8
isort